Replace Content Component to Homepage
