--- src/pages/Homepage/index.css
@@ -1,57 +1,0 @@
-.grid.wide{
-    overflow: hidden;
-    overflow: hidden;
-    min-height: 100vh;
-    padding-top: 3rem;
-    padding-bottom: 3rem;
-}
-
-.preview.col{
-    background-color: transparent;
-    background-clip: content-box;
-    text-align: center;        
-    margin-top: 0.8rem;
-    margin-bottom:  0.8rem;
-}
-
-.root2{
-    min-height: 100vh;
-    padding-top: 12rem;
-    padding-bottom: 15rem;
-}
-.preview{
-    width: 40rem;
-    height: 40rem;
-    position: relative;
-}
-.preview__image{
-    border-radius: 3rem;
-    width: 100%;
-    height: 100%;
-    object-fit: cover;
-    overflow:hidden;
-    scale: 0.5;
-    margin: 0 auto;
-}
-
-.preview .preview__title{
-    position: absolute;
-    white-space: nowrap;
-    bottom: 1.4rem;
-    left:50%;
-    transform: translateX(-50%);
-    text-transform: capitalize;
-    color: var(--color-white);
-    transition-duration: 0.3s
-}
-
-.preview:hover .preview__title{
-    color: var(--primary-color)
-
-}
-
-
-.preview:hover img{
-    opacity: 0.85;
-    border: solid 0.2rem var(--primary-color);
-}
