+++ src/pages/Homepage/index.scss
@@ -18,6 +18,7 @@
     }
 
     .preview__title {
+      border-top: none !important;
       position: absolute;
       white-space: nowrap;
       bottom: 1.4rem;
